+++ poetry/console/commands/check.py
@@ -10,6 +10,16 @@ class CheckCommand(Command):
 
     def handle(self):
         # Load poetry and display errors, if any
-        self.poetry.check(self.poetry.local_config, strict=True)
+        check_result = self.poetry.check(self.poetry.local_config, strict=True)
+        if not check_result["errors"] and not check_result["warnings"]:
+            self.info("All set!")
 
-        self.info("All set!")
+            return 0
+
+        for error in check_result["errors"]:
+            self.error("Error: {}".format(error))
+
+        for error in check_result["warnings"]:
+            self.line("<warning>Warning: {}</warning>".format(error))
+
+        return 1